Renamed MembersListInterface
