+++ src/Ownership/Domain/Entity/MembersList.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Ownership\Domain\Entity;
 
-use App\Twitter\Domain\Publication\MembersListInterface;
+use App\Ownership\Domain\Entity\MembersListInterface;
 use DateTime;
 use DateTimeInterface;
 use DateTimeZone;
