+++ src/Ownership/Domain/Entity/MembersListInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Twitter\Domain\Publication;
+namespace App\Ownership\Domain\Entity;
 
 interface MembersListInterface
 {
