+++ src/Twitter/Domain/Curation/Entity/StatusTrait.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Twitter\Domain\Curation\Entity;
 
-use App\Twitter\Domain\Publication\MembersListInterface;
+use App\Ownership\Domain\Entity\MembersListInterface;
 use App\Twitter\Domain\Publication\StatusInterface;
 use DateTimeInterface;
 use Doctrine\Common\Collections\ArrayCollection;
