+++ src/Twitter/Domain/Publication/Exception/LockedPublishersListException.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Twitter\Domain\Publication\Exception;
 
-use App\Twitter\Domain\Publication\MembersListInterface;
+use App\Ownership\Domain\Entity\MembersListInterface;
 use RuntimeException;
 
 class LockedPublishersListException extends RuntimeException
