+++ src/Twitter/Domain/Publication/Repository/TaggedStatusRepositoryInterface.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Twitter\Domain\Publication\Repository;
 
-use App\Twitter\Domain\Publication\MembersListInterface;
+use App\Ownership\Domain\Entity\MembersListInterface;
 use App\Twitter\Domain\Publication\StatusInterface;
 
 interface TaggedStatusRepositoryInterface
