+++ src/Twitter/Domain/Publication/StatusInterface.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Twitter\Domain\Publication;
 
-use App\Twitter\Domain\Publication\MembersListInterface;
+use App\Ownership\Domain\Entity\MembersListInterface;
 use DateTimeInterface;
 use Doctrine\Common\Collections\Collection;
