+++ src/Twitter/Domain/Publication/Tag.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace App\Twitter\Domain\Publication;
 
 use App\Ownership\Domain\Entity\MembersList;
+use App\Ownership\Domain\Entity\MembersListInterface;
 
 class Tag
 {
