+++ src/Twitter/Domain/Publication/TaggedStatus.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace App\Twitter\Domain\Publication;
 
 use App\Ownership\Domain\Entity\MembersList;
+use App\Ownership\Domain\Entity\MembersListInterface;
 use App\Twitter\Infrastructure\Http\Entity\Status;
 use App\Twitter\Domain\Publication\Exception\InvalidTagPropertyException;
 use DateTimeInterface;
