+++ src/Twitter/Infrastructure/Http/Entity/Status.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Twitter\Infrastructure\Http\Entity;
 
-use App\Twitter\Domain\Publication\MembersListInterface;
+use App\Ownership\Domain\Entity\MembersListInterface;
 use App\Twitter\Domain\Publication\StatusInterface;
 use App\Twitter\Domain\Curation\Entity\StatusTrait;
 use DateTimeInterface;
