+++ src/Twitter/Infrastructure/Publication/Persistence/StatusPersistenceInterface.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Twitter\Infrastructure\Publication\Persistence;
 
-use App\Twitter\Domain\Publication\MembersListInterface;
+use App\Ownership\Domain\Entity\MembersListInterface;
 use App\Twitter\Domain\Publication\StatusInterface;
 use App\Twitter\Infrastructure\Http\AccessToken\AccessToken;
 use Doctrine\ORM\EntityManagerInterface;
